pyproject.toml
[project]
name = "saas-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "colorlog>=6.9.0",
    "django>=5.2.5",
    "django-ratelimit>=4.1.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular[sidecar]>=0.28.0",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.10",
    "python-decouple>=3.8",
]
